+++ command/test/command/help_command_test.ts
@@ -1,9 +1,7 @@
-import { assertEquals, gt, lt, stripColor } from "../../../dev_deps.ts";
+import { assertEquals } from "../../../dev_deps.ts";
 import { CompletionsCommand } from "../../completions/mod.ts";
 import { HelpCommand } from "../../help/mod.ts";
 import { Command } from "../../command.ts";
-
-const isNightly: boolean = Deno.version.deno.match(/nightly/) !== null;
 
 function command(defaultOptions?: boolean, hintOption?: boolean) {
   const cmd = new Command()
@@ -13,6 +11,7 @@ function command(defaultOptions?: boolean, hintOption?: boolean) {
     .help({
       hints: true,
       types: true,
+      colors: false,
     });
 
   if (!defaultOptions) {
@@ -68,96 +67,12 @@ function command(defaultOptions?: boolean, hintOption?: boolean) {
 }
 
 Deno.test({
-  name: "command: help command with line break (deno <= v1.4.0)",
-  ignore: isNightly || gt(Deno.version.deno, "1.4.0"),
-  fn() {
-    const output: string = command(true, false).getHelp();
-
-    assertEquals(
-      stripColor(output),
-      `
-  Usage:   COMMAND
-  Version: v1.0.0 
-
-  Description:
-
-    Test description ...
-
-  Options:
-
-    -h, --help                     - Show this help.                                               
-    -V, --version                  - Show the version number for this program.                     
-    -t, --test       [val:string]  - test description                                              
-    -D, --default    [val:string]  - I have a default value!                    (Default: test)    
-    -r, --required   [val:string]  - I am required!                             (required)         
-    -d, --depends    [val:string]  - I depend on test!                          (Depends: --test)  
-    -c, --conflicts  [val:string]  - I conflict with test!                      (Conflicts: --test)
-
-  Commands:
-
-    help         [command:command]               - Show this help or the help of a sub-command.
-    completions                                  - Generate shell completions.                 
-    sub-command  <input:string> <output:string>  - sub command description.                    
-
-  Environment variables:
-
-    SOME_ENV_VAR    <value:number>  - Description ...  
-    SOME_ENV_VAR_2  <value:string>  - Description 2 ...
-
-`,
-    );
-  },
-});
-
-Deno.test({
-  name:
-    "command: help command with line break but without default options (deno <= v1.4.0)",
-  ignore: isNightly || gt(Deno.version.deno, "1.4.0"),
-  fn() {
-    const output: string = command(false, false).getHelp();
-
-    assertEquals(
-      stripColor(output),
-      `
-  Usage:   COMMAND
-  Version: v1.0.0 
-
-  Description:
-
-    Test description ...
-
-  Options:
-
-    -t, --test       [val:string]  - test description                            
-    -D, --default    [val:string]  - I have a default value!  (Default: test)    
-    -r, --required   [val:string]  - I am required!           (required)         
-    -d, --depends    [val:string]  - I depend on test!        (Depends: --test)  
-    -c, --conflicts  [val:string]  - I conflict with test!    (Conflicts: --test)
-
-  Commands:
-
-    help         [command:command]               - Show this help or the help of a sub-command.
-    completions                                  - Generate shell completions.                 
-    sub-command  <input:string> <output:string>  - sub command description.                    
-
-  Environment variables:
-
-    SOME_ENV_VAR    <value:number>  - Description ...  
-    SOME_ENV_VAR_2  <value:string>  - Description 2 ...
-
-`,
-    );
-  },
-});
-
-Deno.test({
-  name: "command: help command with line break (deno >= v1.4.1)",
-  ignore: !isNightly && lt(Deno.version.deno, "1.4.1"),
+  name: "command: help command with line break",
   fn() {
     const output: string = command(true, true).getHelp();
 
     assertEquals(
-      stripColor(output),
+      output,
       `
   Usage:   COMMAND
   Version: v1.0.0 
@@ -195,14 +110,12 @@ Deno.test({
 });
 
 Deno.test({
-  name:
-    "command: help command with line break but without default options (deno >= v1.4.1)",
-  ignore: !isNightly && lt(Deno.version.deno, "1.4.1"),
+  name: "command: help command with line break but without default options",
   fn() {
     const output: string = command(false, true).getHelp();
 
     assertEquals(
-      stripColor(output),
+      output,
       `
   Usage:   COMMAND
   Version: v1.0.0