+++ examples/command/help_option_and_command.ts
@@ -11,6 +11,7 @@ await new Command()
   .help({
     types: true,
     hints: true,
+    colors: true,
   })
   .env(
     "EXAMPLE_ENVIRONMENT_VARIABLE=<value:boolean>",